Larissa id#12601

See merge request escoladeti2018time03/minder-api!27

+++ src/main/java/br/minder/telefone/TelefoneService.java
@@ -6,7 +6,6 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import br.minder.telefone.comandos.BuscarTelefone;
 import br.minder.telefone.comandos.CriarTelefone;
 import br.minder.telefone.comandos.EditarTelefone;
@@ -18,8 +17,11 @@ public class TelefoneService {
 	private TelefoneRepository repo;
 
 	public Optional<TelefoneId> salvar(CriarTelefone comando) {
-		Telefone novo = repo.save(new Telefone(comando));
-		return Optional.of(novo.getId());
+		if (comando.getDdd() != 0 && comando.getNumero() != 0) {
+			Telefone novo = repo.save(new Telefone(comando));
+			return Optional.of(novo.getId());
+		}
+		return Optional.empty();
 	}
 
 	public Optional<BuscarTelefone> encontrar(TelefoneId id) {
@@ -34,16 +36,19 @@ public class TelefoneService {
 	public Optional<List<BuscarTelefone>> encontrar() {
 		List<BuscarTelefone> resultados = new ArrayList<>();
 		List<Telefone> telefones = repo.findAll();
-		for (Telefone telefone : telefones) {
-			BuscarTelefone tel = new BuscarTelefone(telefone);
-			resultados.add(tel);
+		if (!telefones.isEmpty()) {
+			for (Telefone telefone : telefones) {
+				BuscarTelefone tel = new BuscarTelefone(telefone);
+				resultados.add(tel);
+			}
+			return Optional.of(resultados);
 		}
-		return Optional.of(resultados);
+		return Optional.empty();
 	}
 
 	public Optional<TelefoneId> alterar(EditarTelefone comando) {
 		Optional<Telefone> optional = repo.findById(comando.getId());
-		if (optional.isPresent()) {
+		if (comando.getDdd() != 0 && comando.getNumero() != 0 && optional.isPresent()) {
 			Telefone tel = optional.get();
 			tel.apply(comando);
 			repo.save(tel);
@@ -53,7 +58,10 @@ public class TelefoneService {
 	}
 
 	public Optional<String> deletar(TelefoneId idTelefone) {
-		repo.deleteById(idTelefone);
-		return Optional.of("Telefone ===> " + idTelefone + ": deletado com sucesso");
+		if (repo.findById(idTelefone).isPresent()) {
+			repo.deleteById(idTelefone);
+			return Optional.of("Telefone ===> " + idTelefone + ": deletado com sucesso");
+		}
+		return Optional.empty();
 	}
 }
